Remove unnecessary (un)boxing

+++ client/src/main/java/org/evosuite/assertion/CompareAssertion.java
@@ -115,7 +115,7 @@ public class CompareAssertion extends Assertion {
 					return true; // TODO - true or false?
 			else {
 				try {
-					return comparable.compareTo(dest.getObject(scope)) == ((Integer) value).intValue();
+					return comparable.compareTo(dest.getObject(scope)) == (Integer) value;
 				} catch (Exception e) {
 					return true; // TODO - true or false?
 				}
