+++ client/src/main/java/org/evosuite/symbolic/expr/DistanceCalculator.java
@@ -115,7 +115,7 @@ public class DistanceCalculator implements ConstraintVisitor<Object, Void> {
 			IntegerConstant right_constant = (IntegerConstant) n.getRightOperand();
 			StringUnaryToIntegerExpression left_string_expr = (StringUnaryToIntegerExpression) n.getLeftOperand();
 
-			if (right_constant.getConcreteValue().longValue() != 0L) {
+			if (right_constant.getConcreteValue() != 0L) {
 				return -1;
 			}
 
@@ -258,7 +258,7 @@ public class DistanceCalculator implements ConstraintVisitor<Object, Void> {
 			if (((IntegerUnaryExpression) n.getLeftOperand()).getOperator() == Operator.ISDIGIT) {
 				Long leftObject = (Long) ((IntegerUnaryExpression) n.getLeftOperand()).getOperand().accept(exprExecutor,
 						null);
-				long left_operand = leftObject.longValue();
+				long left_operand = leftObject;
 				char theChar = (char) left_operand;
 				if ((n.getComparator() == Comparator.EQ && rightVal == 1L)
 						|| (n.getComparator() == Comparator.NE && rightVal == 0L)) {
@@ -279,7 +279,7 @@ public class DistanceCalculator implements ConstraintVisitor<Object, Void> {
 			} else if (((IntegerUnaryExpression) n.getLeftOperand()).getOperator() == Operator.ISLETTER) {
 				Long leftObject = (Long) ((IntegerUnaryExpression) n.getLeftOperand()).getOperand().accept(exprExecutor,
 						null);
-				long left_operand = leftObject.longValue();
+				long left_operand = leftObject;
 				char theChar = (char) left_operand;
 				if ((n.getComparator() == Comparator.EQ && rightVal == 1L)
 						|| (n.getComparator() == Comparator.NE && rightVal == 0L)) {
