+++ client/src/test/java/com/examples/with/different/packagename/concolic/TestCase100.java
@@ -25,7 +25,7 @@ public class TestCase100 {
 		int intValue = int0;
 		try {
 			Integer i = Integer.parseInt(string0);
-			intValue = i.intValue();
+			intValue = i;
 		} catch (NumberFormatException ex) {
 			if (intValue == -1) {
 				return true;
