+++ client/src/test/java/com/examples/with/different/packagename/concolic/TestCase99.java
@@ -23,7 +23,7 @@ public class TestCase99 {
 
 	public static boolean test(String myStr) {
 		Integer i = Integer.parseInt(myStr);
-		int intValue = i.intValue();
+		int intValue = i;
 		if (intValue == 10) {
 			return true;
 		} else {
