+++ client/src/test/java/org/evosuite/symbolic/solver/TestSolverSimpleMath.java
@@ -338,7 +338,7 @@ public abstract class TestSolverSimpleMath extends TestSolver {
 		assertNotNull(solution);
 		Double var0 = (Double) solution.get("var0");
 
-		assertTrue(var0.doubleValue() != 0);
+		assertTrue(var0 != 0);
 		assertTrue(var0.intValue() == 0);
 	}
