+++ master/src/test/java/com/examples/with/different/packagename/test/PolyExample.java
@@ -23,7 +23,7 @@ public class PolyExample {
 	public void testMe(Object o) {
 		if (o instanceof Integer) {
 			Integer i = (Integer) o;
-			if (i.intValue() == 17) {
+			if (i == 17) {
 				System.out.println("test");
 			}
 		}
